requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
idna==3.4
isort==5.10.1
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
psycopg2==2.9.5
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
